backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.4
pandas==2.1.4
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
